# src/main/resources/openapi.yaml
openapi: 3.0.3
tags:
  - Delivery
info:
  title: Delivery Route & Driver Assignment API
  description: API for managing delivery orders and driver assignments.
  version: 1.0.0

paths:
  /orders:
    post:
      summary: Create a new order and assign a driver
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

components:
  schemas:
    CreateOrderRequest:
      type: object
      properties:
        restaurantLatitude:
          type: number
          format: double
          example: 12.9716
        restaurantLongitude:
          type: number
          format: double
          example: 77.5946
        customerLatitude:
          type: number
          format: double
          example: 12.9800
        customerLongitude:
          type: number
          format: double
          example: 77.6000

    DriverResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        available:
          type: boolean

    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        restaurantLatitude:
          type: number
          format: double
        restaurantLongitude:
          type: number
          format: double
        customerLatitude:
          type: number
          format: double
        customerLongitude:
          type: number
          format: double
        status:
          type: string
          example: ASSIGNED
        assignedDriver:
          $ref: '#/components/schemas/DriverResponse'